+++ go/src/github.com/bitrise-core/bitrise-init/_tests/integration/manual_config_test.go
@@ -8,28 +8,27 @@ import (
 
 	"os"
 
-	"github.com/bitrise-io/go-utils/cmdex"
+	"github.com/bitrise-core/bitrise-init/models"
+	"github.com/bitrise-core/bitrise-init/steps"
+	"github.com/bitrise-io/go-utils/command"
 	"github.com/bitrise-io/go-utils/fileutil"
 	"github.com/bitrise-io/go-utils/pathutil"
 	"github.com/stretchr/testify/require"
-
-	"github.com/bitrise-core/bitrise-init/models"
-	"github.com/bitrise-core/bitrise-init/steps"
 )
 
 func TestManualConfig(t *testing.T) {
 	tmpDir, err := pathutil.NormalizedOSTempDirPath("__manual-config__")
 	require.NoError(t, err)
-	// defer func() {
-	// 	require.NoError(t, os.RemoveAll(tmpDir))
-	// }()
+	defer func() {
+		require.NoError(t, os.RemoveAll(tmpDir))
+	}()
 
 	t.Log("manual-config")
 	{
 		manualConfigDir := filepath.Join(tmpDir, "manual-config")
 		require.NoError(t, os.MkdirAll(manualConfigDir, 0777))
 
-		cmd := cmdex.NewCommand(binPath(), "--ci", "manual-config", "--output-dir", manualConfigDir)
+		cmd := command.New(binPath(), "--ci", "manual-config", "--output-dir", manualConfigDir)
 		out, err := cmd.RunAndReturnTrimmedCombinedOutput()
 		require.NoError(t, err, out)
 
@@ -47,15 +46,9 @@ var customConfigVersions = []interface{}{
 	steps.ActivateSSHKeyVersion,
 	steps.GitCloneVersion,
 	steps.ScriptVersion,
-	steps.ScriptVersion,
+	steps.InstallMissingAndroidToolsVersion,
 	steps.GradleRunnerVersion,
 	steps.DeployToBitriseIoVersion,
-
-	// custom
-	models.FormatVersion,
-	steps.ActivateSSHKeyVersion,
-	steps.GitCloneVersion,
-	steps.ScriptVersion,
 
 	// fastlane
 	models.FormatVersion,
@@ -107,6 +100,12 @@ var customConfigVersions = []interface{}{
 	steps.RecreateUserSchemesVersion,
 	steps.XcodeTestMacVersion,
 	steps.DeployToBitriseIoVersion,
+
+	// other
+	models.FormatVersion,
+	steps.ActivateSSHKeyVersion,
+	steps.GitCloneVersion,
+	steps.ScriptVersion,
 
 	// xamarin
 	models.FormatVersion,
@@ -198,41 +197,13 @@ configs:
           - git-clone@%s: {}
           - script@%s:
               title: Do anything with Script step
-          - script@%s:
-              title: Update Android Extra packages
-              inputs:
-              - content: |
-                  #!/bin/bash
-                  set -ex
-
-                  echo y | android update sdk --no-ui --all --filter platform-tools | grep 'package installed'
-
-                  echo y | android update sdk --no-ui --all --filter extra-android-m2repository | grep 'package installed'
-                  echo y | android update sdk --no-ui --all --filter extra-google-m2repository | grep 'package installed'
-                  echo y | android update sdk --no-ui --all --filter extra-google-google_play_services | grep 'package installed'
+          - install-missing-android-tools@%s: {}
           - gradle-runner@%s:
               inputs:
               - gradle_file: $GRADLE_BUILD_FILE_PATH
               - gradle_task: $GRADLE_TASK
               - gradlew_path: $GRADLEW_PATH
           - deploy-to-bitrise-io@%s: {}
-  custom:
-    custom-config: |
-      format_version: %s
-      default_step_lib_source: https://github.com/bitrise-io/bitrise-steplib.git
-      trigger_map:
-      - push_branch: '*'
-        workflow: primary
-      - pull_request_source_branch: '*'
-        workflow: primary
-      workflows:
-        primary:
-          steps:
-          - activate-ssh-key@%s:
-              run_if: '{{getenv "SSH_RSA_PRIVATE_KEY" | ne ""}}'
-          - git-clone@%s: {}
-          - script@%s:
-              title: Do anything with Script step
   fastlane:
     default-fastlane-config: |
       format_version: %s
@@ -356,6 +327,23 @@ configs:
               - project_path: $BITRISE_PROJECT_PATH
               - scheme: $BITRISE_SCHEME
           - deploy-to-bitrise-io@%s: {}
+  other:
+    other-config: |
+      format_version: %s
+      default_step_lib_source: https://github.com/bitrise-io/bitrise-steplib.git
+      trigger_map:
+      - push_branch: '*'
+        workflow: primary
+      - pull_request_source_branch: '*'
+        workflow: primary
+      workflows:
+        primary:
+          steps:
+          - activate-ssh-key@%s:
+              run_if: '{{getenv "SSH_RSA_PRIVATE_KEY" | ne ""}}'
+          - git-clone@%s: {}
+          - script@%s:
+              title: Do anything with Script step
   xamarin:
     default-xamarin-config: |
       format_version: %s