+++ go/src/github.com/bitrise-core/bitrise-init/utility/podfile.go
@@ -8,8 +8,8 @@ import (
 
 	"encoding/json"
 
-	"github.com/bitrise-io/depman/pathutil"
 	"github.com/bitrise-io/go-utils/fileutil"
+	"github.com/bitrise-io/go-utils/pathutil"
 	"github.com/bitrise-tools/go-xcode/xcodeproj"
 )